% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_aliases_1in1000.R
\docType{data}
\name{colour_aliases_1in1000}
\alias{colour_aliases_1in1000}
\alias{color_aliases_1in1000}
\title{Named colour vector associated with 1 in 1000}
\format{
An object of class \code{character} of length 42.
}
\usage{
colour_aliases_1in1000
}
\description{
This vector contains names associated with different colours from 1 in 1000
palette.
}
\examples{
colour_aliases_1in1000

colour_aliases_1in1000[c("high carbon exposure", "low carbon exposure")]
}
\keyword{datasets}
