% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_aliases_2dii.R
\docType{data}
\name{colour_aliases_2dii}
\alias{colour_aliases_2dii}
\alias{color_aliases_2dii}
\title{Named colour vector associated with 2DII values}
\format{
An object of class \code{character} of length 37.
}
\usage{
colour_aliases_2dii
}
\description{
This vector contains names associated with different colours from 2DII
palette.
}
\examples{
colour_aliases_2dii

colour_aliases_2dii[c("equity", "bonds")]
}
\keyword{datasets}
