% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_aliases_pacta.R
\docType{data}
\name{colour_aliases_pacta}
\alias{colour_aliases_pacta}
\alias{color_aliases_pacta}
\title{Named colour vector associated with PACTA sectors and technologies}
\format{
An object of class \code{character} of length 67.
}
\usage{
colour_aliases_pacta
}
\description{
This vector contains named colours associated with PACTA sectors and
technologies.
}
\examples{
colour_aliases_pacta

colour_aliases_pacta[c("ice", "hybrid", "electric")]

colour_aliases_pacta[c("steel", "aviation", "cement")]
}
\keyword{datasets}
