% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_aliases_survey.R
\docType{data}
\name{colour_aliases_survey}
\alias{colour_aliases_survey}
\alias{color_aliases_survey}
\title{Named colour vector associated with user survey}
\format{
An object of class \code{character} of length 10.
}
\usage{
colour_aliases_survey
}
\description{
This vector contains named colours associated with user survey results plots.
}
\examples{
colour_aliases_survey

colour_aliases_survey[c("climate_strategy_concrete_goals")]

colour_aliases_pacta[c("Concrete goals", "Climate strategy")]
}
\keyword{datasets}
