% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colours.R
\name{get_colours}
\alias{get_colours}
\title{Retrieve colour hex values from a palette}
\usage{
get_colours(
  colour_names = "red",
  palette = r2dii.colours::palette.1in1000.plot
)
}
\arguments{
\item{colour_names}{Character string or a vector of character strings with
names of colours for which you want to retrieve the hex values.}

\item{palette}{The dataframe from which the hex values are retrieved. It
should contain columns \code{label} and \code{hex}.}
}
\value{
A vector of strings with hex codes.
}
\description{
Retrieve colour hex values from a palette
}
\examples{
# use default palette
get_colours(c("red", "blue"))

# specify which palette to use
get_colours(c("red", "green"), palette = palette.1in1000.goodbad)
}
