% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.colours-package.R
\docType{package}
\name{r2dii.colours-package}
\alias{r2dii.colours}
\alias{r2dii.colours-package}
\title{r2dii.colours: 2 Degrees Investing Colour Palettes in Different Formats}
\description{
Get colour values from different colour palettes used by 2 Degrees Investing (2DII) organization in their reserach streams. Different ways to obtain the colour values are available: dataframe or a function call.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/r2dii.colours}
  \item \url{https://2degreesinvesting.github.io/r2dii.colours/}
}

}
\author{
\strong{Maintainer}: Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID}) [contractor]

}
\keyword{internal}
