% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_1in1000.R
\name{scale_colour_1in1000}
\alias{scale_colour_1in1000}
\alias{scale_color_1in1000}
\alias{scale_fill_1in1000}
\title{Custom 1 in 1000 colour and fill scales}
\usage{
scale_colour_1in1000(colour_groups = NULL, ...)

scale_fill_1in1000(colour_groups = NULL, ...)
}
\arguments{
\item{colour_groups}{A vector containing groups variable to which colours are
assigned. It is needed when the data assigned to \code{colour} aesthetic are
not all contained in colour aliases of the palette.}

\item{...}{Other parameters passed on to \code{ggplot2::discrete_scale()}.}
}
\value{
An object of class "ScaleDiscrete".
}
\description{
A custom discrete colour and fill scales with colours from 1 in 1000 palette.
}
\examples{
library(ggplot2, warn.conflicts = FALSE)
library(r2dii.plot, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)

ggplot(sda \%>\% filter(emission_factor_metric == 'projected')) +
 geom_line(aes(x = year, y = emission_factor_value, colour = sector)) +
 scale_colour_1in1000()

ggplot(mpg) +
  geom_histogram(aes(cyl, fill = class), position = "dodge", bins = 5) +
  scale_fill_1in1000()
}
\seealso{
Other r2dii scales: 
\code{\link{scale_colour_2dii}()}
}
\concept{r2dii scales}
