% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\docType{data}
\name{colour_palettes}
\alias{colour_palettes}
\alias{palette_1in1000_background}
\alias{palette_1in1000_goodbad}
\alias{palette_1in1000_plot}
\alias{palette_2dii_plot}
\alias{palette_2dii_scenario}
\alias{palette_2dii_sector}
\alias{palette_2dii_power}
\alias{palette_2dii_automotive}
\alias{palette_2dii_oil_gas}
\alias{palette_2dii_fossil_fuels}
\title{Colour palette datasets}
\format{
An object of class \code{character} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.
}
\usage{
palette_1in1000_background

palette_1in1000_goodbad

palette_1in1000_plot

palette_2dii_plot

palette_2dii_scenario

palette_2dii_sector

palette_2dii_power

palette_2dii_automotive

palette_2dii_oil_gas

palette_2dii_fossil_fuels
}
\description{
All datasets have at least two columns:
\itemize{
\item \code{label}: Text label of the colour.
\item \code{hex}: Hex code of the colour.
}
}
\examples{
palette_1in1000_background

palette_1in1000_goodbad

palette_1in1000_plot
}
\keyword{datasets}
