% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\docType{data}
\name{colour_palettes}
\alias{colour_palettes}
\alias{palette.1in1000.background}
\alias{palette.1in1000.goodbad}
\alias{palette.1in1000.plot}
\alias{palette_1in1000_background}
\alias{palette_1in1000_goodbad}
\alias{palette_1in1000_plot}
\title{Colour palette datasets}
\format{
An object of class \code{character} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.

An object of class \code{character} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.
}
\usage{
palette.1in1000.background

palette.1in1000.goodbad

palette.1in1000.plot

palette_1in1000_background

palette_1in1000_goodbad

palette_1in1000_plot
}
\description{
All datasets have at least two columns:
\itemize{
\item \code{label}: Text label of the colour.
\item \code{hex}: Hex code of the colour.
}
}
\examples{
palette_1in1000_background

palette_1in1000_goodbad

palette_1in1000_plot
}
\keyword{datasets}
